Fix for External API
- RAML
- Resources
- Clients

Also includes fix on Units collectiion where _uds is sometimes a map, sometimes a list, to a list (this has to be done also on IT18).
Note: we will have to decide if this should be an array or a map (originally, it should be a map).

- Access External
- Ingest External

Fix AbstractCommonClient.staticConsumeAnyEntityAndClose

Add some FIXME for IT18