Merge branch 'bugs_13211' into 'develop'
Bugs #13211: change error message when trying to delete an internal schema item

Currently we reject the removal of an internal schema with a "not found" error message. Here we change the error message to indicate that the removal is not possible because the path corresponds to an internal schema.

See merge request vitam/vitam!10194

(cherry picked from commit eef0e81dac4f413022ed070ef4a06798c9a8e549)

9584f9b6 Bugs #13211: change error message when trying to delete an internal schema item

Co-authored-by: Etienne Bernard <ebernard@ippon.fr>